move spec file to data as well
somehow this is a huge fuckup and doesn't make any sense to me.
copr should only be building a provided SPRM, but it requires the source
as well. beside the spec file must be under the same dir as the rest of
the files.

it would be way better to simply use the rpmbuild default directories,
e.g. SOURCES or SRPM